requirements.txt
submitit
torchmetrics[image]
torchvision
transformers
